+++ sdk/@launchdarkly/observability-python/ldobserve/_otel/configuration.py
@@ -48,40 +48,6 @@ _MAX_EXPORT_BATCH_SIZE = 128 * 1024
 _MAX_QUEUE_SIZE = 1024 * 1024
 
 
-def _extend_disabled_instrumentations(*instrumentations: str) -> None:
-    """
-    Extend the OTEL_PYTHON_DISABLED_INSTRUMENTATIONS environment variable with additional instrumentations.
-
-    This function reads the current value of OTEL_PYTHON_DISABLED_INSTRUMENTATIONS,
-    adds the provided instrumentations to the comma-separated list, and sets the
-    environment variable with the extended list.
-
-    Args:
-        *instrumentations: Variable number of instrumentation names to add to the disabled list.
-                          These should be the names of OpenTelemetry instrumentations to disable.
-
-    Example:
-        >>> extend_disabled_instrumentations("redis", "kafka")
-        # If OTEL_PYTHON_DISABLED_INSTRUMENTATIONS was "grpc_client",
-        # it will become "grpc_client,redis,kafka"
-    """
-    current_value = os.getenv(OTEL_PYTHON_DISABLED_INSTRUMENTATIONS, "")
-
-    # Split the current value by comma and strip whitespace
-    current_list = [item.strip() for item in current_value.split(",") if item.strip()]
-
-    # Add new instrumentations, avoiding duplicates
-    for instrumentation in instrumentations:
-        if instrumentation.strip() and instrumentation.strip() not in current_list:
-            current_list.append(instrumentation.strip())
-
-    # Join the list back into a comma-separated string
-    new_value = ",".join(current_list)
-
-    # Set the environment variable
-    os.environ[OTEL_PYTHON_DISABLED_INSTRUMENTATIONS] = new_value
-
-
 def _build_resource(
     project_id: str,
     service_name: typing.Optional[str],
@@ -139,9 +105,6 @@ class _OTELConfiguration:
             service_version=self._config.service_version,
             environment=self._config.environment,
         )
-
-        if self._config.disabled_instrumentations:
-            _extend_disabled_instrumentations(*self._config.disabled_instrumentations)
 
         sampler = CustomSampler()
         # TODO: Get and set config.
@@ -204,10 +167,6 @@ class _OTELConfiguration:
             logging.basicConfig(level=self._config.log_level)
             logging.getLogger().addHandler(self._log_handler)
 
-        # Instruct auto-instrumentation to not instrument logging.
-        # We will either have already done it, or it is disabled.
-        _extend_disabled_instrumentations("logging")
-
         metric_reader = PeriodicExportingMetricReader(
             exporter=OTLPMetricExporter(
                 f"{self._config.otlp_endpoint}/v1/metrics",
@@ -232,3 +191,6 @@ class _OTELConfiguration:
         )
         metrics.set_meter_provider(self._meter_provider)
         self._meter = self._meter_provider.get_meter(__name__)
+
+    def activate(self) -> None:
+        pass
